Mtr profiling setup
A complete generated table fixture was added. Unfortunately, imported
tables behave bad in innodb. They are subject for checksumming on and on,
so it's better to ALTER TABLE FORCE them before actually profiling.

Usage:
Put vehicle_reg.ibd in mysq-test/std_data. (cant push it; too large)
./mtr --mem main.profile
It will emit notice, that you should rerun it with --start-dirty.
Wait ~two minutes.
Check that var/mysqld.1 directory is there. If not,
it was archived by mtr. Run:
mv var/log/main.profile-innodb/mysqld.1 var

Now rerun with the dirty dir:
./mtr main.profile --start-dirty --vardir=`pwd`/var  --suite=main
You can add --gprof to collect the profiling data.
Make sure you've compiled mariadbd with -DWITH_GPROF=1.
Open gprof.msg and navigate to Call graph.
Dont use IDE to open! The file is huge. Vim/emacs is best, but maybe
sublime would work, too.

+++ sql/records.cc
@@ -326,7 +326,7 @@ bool init_read_record(READ_RECORD *info,THD *thd, TABLE *table,
 
   else if (table->tablesample_method)
   {
-    int method = table->tablesample_method->val_int();
+    longlong method = table->tablesample_method->val_int();
     double fract= (method  % 100) / 100.0;
     info->sample_factor = fract;
     info->sample_counter= (ha_rows)(table->file->records() * fract + 0.5);
